requirements.txt
numpy
torch
transformers
Pillow
pycocotools
hugsvision
torchmetrics
scikit-learn
tqdm
tabulate
torchvision
opencv-python
matplotlib
pandas
setuptools
streamlit
pyqt5